+++ cmd/containers-storage/create.go
@@ -8,7 +8,7 @@ import (
 	"os"
 
 	"github.com/containers/storage"
-	"github.com/containers/storage/opts"
+	"github.com/containers/storage/internal/opts"
 	"github.com/containers/storage/pkg/idtools"
 	"github.com/containers/storage/pkg/mflag"
 	digest "github.com/opencontainers/go-digest"
