+++ cmd/containers-storage/name.go
@@ -6,7 +6,7 @@ import (
 	"os"
 
 	"github.com/containers/storage"
-	"github.com/containers/storage/opts"
+	"github.com/containers/storage/internal/opts"
 	"github.com/containers/storage/pkg/mflag"
 )
