+++ go.mod
@@ -5,7 +5,7 @@ require (
 	github.com/DataDog/zstd v1.4.0 // indirect
 	github.com/Microsoft/go-winio v0.4.12
 	github.com/Microsoft/hcsshim v0.8.6
-	github.com/docker/docker v0.0.0-20171019062838-86f080cff091
+	github.com/docker/docker v0.0.0-20171019062838-86f080cff091 // indirect
 	github.com/docker/go-units v0.4.0
 	github.com/klauspost/compress v1.7.2
 	github.com/klauspost/cpuid v1.2.1 // indirect
